requirements.txt
Flask
langchain
langchain-groq
PyMuPDF
beautifulsoup4
rapidfuzz
python-dotenv
Jinja2